engine/requirements.txt
fastapi
uvicorn
openai
stripe
mixpanel
supabase
python-dotenv
pydantic
httpx
